Modifying the number of workers in tempest
Upstream tempest tests uses ncpu / 2 workers to run tempest.
This is a good pratice to reduce the amount of concurrency failures.
This patch sets the ncpu / 2 using ansible_processor_count variable
in ansible. However, it won't override the number of workers if
tempest_workers is defined, so the default will always be
tempest_workers.

Change-Id: I4486e941bbaaee818dee2d8f383e88fcaef41dff